v 0.12

+++ src/main/java/ru/ketbiev/spring/jproject/controller/UserController.java
@@ -2,6 +2,7 @@ package ru.ketbiev.spring.jproject.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import ru.ketbiev.spring.jproject.exceptionHandling.NoSuchException;
 import ru.ketbiev.spring.jproject.model.Role;
 import ru.ketbiev.spring.jproject.model.User;
 import ru.ketbiev.spring.jproject.service.RoleService;
@@ -10,25 +11,31 @@ import ru.ketbiev.spring.jproject.service.UserService;
 import java.util.*;
 
 @RestController
-@RequestMapping("api")
-public class MainController {
+@RequestMapping("users")
+public class UserController {
 
     @Autowired
     private UserService userService;
+
     @Autowired
     private RoleService roleService;
 
-    @GetMapping("/employees")
+    @GetMapping("/users")
     public List<User> showAllEmployees() {
         return userService.getAllUser();
     }
 
-    @GetMapping("/employees/{id}")
-    public Optional<User> getEmployees(@PathVariable Long id) {
+    @GetMapping("/users/{id}")
+    public Optional<User> getEmployees(@PathVariable int id) {
+        User user = userService.getUser(id).orElse(null);
+
+        if (user == null) {
+            throw new NoSuchException("User");
+        }
         return userService.getUser(id);
     }
 
-    @PostMapping("/employees")
+    @PostMapping("/users")
     public User addNewEmployee(@RequestBody User user) {
         Role role = roleService.findByRole("ROLE_USER");
         Set<Role> roles = new HashSet();
@@ -38,14 +45,20 @@ public class MainController {
         return user;
     }
 
-    @PutMapping("/employees")
+    @PutMapping("/users")
     public User updateEmployee (@RequestBody User user) {
         userService.saveUser(user);
         return user;
     }
 
-    @DeleteMapping("/employees/{id}")
-    public String deleteEmployees(@PathVariable Long id) {
+    @DeleteMapping("/users/{id}")
+    public String deleteEmployees(@PathVariable int id) {
+        User user = userService.getUser(id).orElse(null);
+
+        if (user == null) {
+            throw new NoSuchException("User");
+        }
+
         userService.deleteUser(id);
         return "Deleted id = " + id;
     }
